# [WIP] CPU: Refactored code to be generic across vector CPU platforms.