--- a/openmmapi/include/openmm/internal/vectorize_ppc.h
+++ b/openmmapi/include/openmm/internal/vectorize_ppc.h
@@ -343,7 +343,7 @@ static inline ivec4 abs(const ivec4& v) {
     return vec_abs(v.val);
 }
 
-static inline bool any(const __m128i& v) {
+static inline bool any(const ivec4 v) {
     return !vec_all_eq(v, ivec4(0).val);
 }