… openmmapi/include/openmm/internal/vectorize_pnacl.h
… openmmapi/include/openmm/internal/vectorize_sse.h
… platforms/cpu/include/CpuNonbondedForce.h
… platforms/cpu/include/CpuNonbondedForceFvec.h
… platforms/cpu/include/CpuNonbondedForceVec4.h
… platforms/cpu/include/CpuNonbondedForceVec8.h
… platforms/cpu/src/CpuKernels.cpp
… platforms/cpu/src/CpuNonbondedForce.cpp
… platforms/cpu/src/CpuNonbondedForceFvec.cpp
… platforms/cpu/src/CpuNonbondedForceVec8.cpp
… platforms/cpu/src/CpuPlatform.cpp
… tests/TestVectorize.cpp
… tests/TestVectorize8.cpp