@peastman 2020-04-27
Thanks!  I'll try to make a detailed pass through the code soon.  One general comment is that it should follow the code formatting used everywhere else in OpenMM.  Opening braces should always go on the same line, not the next line.  Comments at the top of classes or methods should use the format

```c++
/**
 * This function does something.
 */
```

rather than

```c++
/// This function does something.
```

Parameters should be documented with `@param` rather than `\param`