@mark-mb 2020-04-27
Given my comments in #2658, I'm still concerned about type conversions that give the wrong value:

1) `const auto include = FVEC::expandBitsToMask(~exclusions[i]) & (r2 < cutoffDistance*cutoffDistance);`
2) `const auto t = (r>switchingDistance) & ((r-switchingDistance)*invSwitchingInterval);`

1) works (I think, from the top of my head), because the right part of the & is converted from fvec4 to ivec4 and used as a mask
2) works in SSE / AVX, but it doesn't work in the current implementation of Neon. The left part of the & isn't a mask, but the representation of -1 as float.

(The issue with 2) wasn't present in CpuNonBondedForceVec4, because that was using blend() instead of &)

Some possible solutions:
 - Get comparison operators to return ivec4's and have an ivec4 as input parameter of fvec4 operator &
 - Add a mask type that's only used for these operations and can be casted to fvec and ivec (as in my original patch)
 - ...

Also note that you are including "vectorize_sse.h" directly. That's probably why CI is not very happy.