@peastman 2020-04-27
vectorize_pnacl.h doesn't get tested by CI, since we aren't currently using it anywhere.  I do want to keep it working though.  The name is misleading: it really is just an implementation that uses the LLVM portable vector syntax.  You can test it locally by just replacing vectorize_sse.h with vectorize_pnacl.h.  It should work on all architectures.  This makes it useful for debugging.