@jchodera 2020-04-27
> One general comment is that it should follow the code formatting used everywhere else in OpenMM. 

@peastman : Do you want to add a few sentences about formatting to the [how to contribute guide](https://github.com/openmm/openmm/blob/master/CONTRIBUTING.md)? Seems like we overlooked this!