@dwtowner 2020-04-27
I don't much like the side-effecting way of using & with different types either.

We could just swap both those cases to using blend instead? That's more explicit about intent.

The return type of the comparison operators won't necessarily be an ivec (e.g., AVX-512 returns int for those operators and uses int for blends resulting from those comparisons). Each platform should be free to choose the most appropriate type.

Something like this might work:

struct Simd {
  // Needed to explicit declare mask variables in those places where auto fails.
  using ElementMask = /* Your type here */

  // Needed for converting the exclusion mask into something that can be manipulated in vector elements.
  static ElementMask getMaskFromInt(int);

  // Comparisons return some suitable type
  ElementMask operator<= (Simd, Simd);

  blend(Simd, Simd, ElementMask);

  // Combine masks - not needed if the mask is an int, but might be required for other platform masks.
  // ElementMask operator& (ElementMask, ElementMask)
};