@mark-mb 2020-04-27
> We could just swap both those cases to using blend instead? That's more explicit about intent.

That would be more explicit and clear and less error-prone.

I'm not sure about the performance difference (if any) between calling the SIMD intrinsic blend vs the intrinsic and. Maybe something like "blendWithZero" would allow to optimize this case?

> Something like this might work:
> 
> struct Simd {
> // Needed to explicit declare mask variables in those places where auto fails.
> using ElementMask = /* Your type here */
> 
> // Needed for converting the exclusion mask into something that can be manipulated in vector elements.
> static ElementMask getMaskFromInt(int);
> 
> // Comparisons return some suitable type
> ElementMask operator<= (Simd, Simd);
> 
> blend(Simd, Simd, ElementMask);
> 
> // Combine masks - not needed if the mask is an int, but might be required for other platform masks.
> // ElementMask operator& (ElementMask, ElementMask)
> };

That would also work. Not sure what is best, so I'll let you decide