@peastman 2020-04-27
According to the Intel intrinsics guide, `_mm_and_ps()` has a latency of 1 and througput of 1.  `_mm_blendv_ps()` has a latency of 2 and throughput of 0.7.  So changing that to blend would hurt performance.  I haven't checked other architectures, but it will likely be slower on them too.