@dwtowner 2020-04-28
> 7

Okay, the intrinsics guide is wrong. I've reported that error to get that fixed. For blend typical performance is latency of 1 and throughput of 0.5 (i.e., the same as an AND).