@peastman 2020-04-28
On Windows compilation is failing with this error:

```
D:\a\1\s\openmmapi\include\openmm/internal/vectorize8.h(184): error C2678: binary '&': no operator found which takes a left-hand operand of type '__m256i' (or there is no acceptable conversion)
```

On Mac there's a linker error:

```
[ 68%] Linking CXX shared library ../../../libOpenMMCPU.dylib
Undefined symbols for architecture x86_64:
  "OpenMM::CpuNonbondedForce::NUM_TABLE_POINTS", referenced from:
      OpenMM::CpuNonbondedForce::threadComputeDirect(OpenMM::ThreadPool&, int) in CpuNonbondedForce.cpp.o
      OpenMM::CpuNonbondedForce::exptermsApprox(float) in CpuNonbondedForce.cpp.o
      OpenMM::CpuNonbondedForce::dExptermsApprox(float) in CpuNonbondedForce.cpp.o
      OpenMM::CpuNonbondedForce::erfcApprox(float) in CpuNonbondedForce.cpp.o
ld: symbol(s) not found for architecture x86_64
```

And on ARM it's failing because it's trying to compile the AVX code:

```
In file included from /home/travis/build/openmm/openmm/platforms/cpu/src/CpuNonbondedForceVec8.cpp:26:0:
/home/travis/build/openmm/openmm/openmmapi/include/openmm/internal/vectorize8.h:36:10: fatal error: immintrin.h: No such file or directory
 #include <immintrin.h>
          ^~~~~~~~~~~~~
compilation terminated.
```