@dwtowner 2020-04-28
I've removed the use of & and replaced by an explicit blendZero function as suggested by @mark-mb . On SSE and AVX* the belendZero doesn't make any difference, but it provides a hook for a slightly more optimal code sequence on other platforms if that is possible.

I've addressed all 3 CI build issues, although getting past them may uncover more.

I'm doing local builds with LLVM/NaCl so they shouldn't cause any problems.