@peastman 2020-04-30
It passed when I reran it, so all tests are now looking good!  Is this ready to merge?

I tested this on my desktop machine, which is a fairly typical 4 core, 8 thread i7.  For DHFR, the performance went from 16.16 to 16.72 ns/day, or about a 3.4% improvement.  For ApoA1 it went from 3.73 to 3.81 ns/day, or about a 2.2% improvement.